adapted the dtype for z_release and z_additional in the full3D case

+++ python/benchmark_Full3D.py
@@ -700,9 +700,9 @@ if __name__ == "__main__":
 
     """ Defining the particle set """
     lon_release, lat_release = np.meshgrid(np.linspace(EqPac['minlon'], EqPac['maxlon'], sx), np.linspace(EqPac['minlat'], EqPac['maxlat'], sy))
-    z_release = np.ones(lon_release.shape[0], dtype=np.float32)
+    z_release = np.ones(lon_release.shape[0], dtype=lon_release.dtype)
     lon_additional, lat_additional = np.meshgrid(np.linspace(EqPac['minlon'], EqPac['maxlon'], asx), np.linspace(EqPac['minlat'], EqPac['maxlat'], asy))
-    z_additional = np.ones(lon_additional.shape[0], dtype=np.float32)
+    z_additional = np.ones(lon_additional.shape[0], dtype=lon_additional.dtype)
 
     print("|startlon| = {}, |startlat| = {}, |startdepth| = {}; |lon| = {}; |lat| = {}, |depth| = {}".format(lon_additional.shape[0], lat_additional.shape[0], z_additional.shape[0], lon_release.shape[0], lat_release.shape[0], z_release.shape[0]))
     pset = ParticleSet.from_list(fieldset=fieldset,         # the fields on which the particles are advected
